Overhaul initialization
Dispense with the literate org-mode-based setup in favor or raw
Lisp. Biggest change is use of raw `package.el` setup instead of
`el-get` so I'll finally learn how Emacs really works :P

Overall structure is based on Steve Purcell's setup (https://github.com/purcell/emacs.d).

Migration isn't totally finished, but a lot of the old functionality is
present (and a number of new features and commands are in place).